fix unlike function and add livewire like component to preview page

+++ app/Http/Livewire/PlaceLikeComponent.php
@@ -11,15 +11,15 @@ class PlaceLikeComponent extends Component
     public $like;
     public $likes;
     public $place;
-    
+
     public function mount($place)
     {
         $this->like = $place->likedBy(Auth::user());
         $this->likes = $place->likes->count();
         $this->place = $place;
     }
-    
-    
+
+
     public function render()
     {
         return view('livewire.place-like-component');
@@ -33,7 +33,7 @@ class PlaceLikeComponent extends Component
             $place_ = Place::where('id', $this->place->id)->get()->first();
             // if liked by user, remove from table
             if ($place_->likedBy(Auth::user())) {
-                $query = $this->place->likes()->where('place_id', $this->place->id)->delete();
+                $query = $this->place->likes()->where('place_id', $this->place->id)->where('user_id', Auth::user()->id)->delete();
             } else {
                 $query = $this->place->likes()->create([
                     'user_id' => Auth::user()->id,
@@ -46,7 +46,7 @@ class PlaceLikeComponent extends Component
                 $this->likes = $place_->likes->count();
             }
         } else {
-            redirect('/login');
+            return redirect()->guest('login');
         }
     }
 }
